requirements.txt
amqp==5.1.1

asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-computed-property==0.3.0
django-phonenumber-field==6.1.0
greenlet==1.1.2
kombu==5.2.4
packaging==21.3
phonenumbers==8.12.47
Pillow==9.1.0
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
